Call reconcile again when files changed
